combine datafusion_ext and datafusion_planner
