+++ crates/datafusion_ext/Cargo.toml
@@ -6,10 +6,23 @@ edition = { workspace = true }
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+default = ["unicode_expressions"]
+unicode_expressions = []
+
 [dependencies]
 datafusion = { workspace = true }
+metastore_client = { path = "../metastore_client" }
+async-trait = "0.1.72"
+async-recursion = "1.0.4"
 uuid = { version = "1.4.0", features = ["v4", "fast-rng", "macro-diagnostics"] }
 regex = "1.8"
 once_cell = "1.18.0"
 tracing = "0.1"
 thiserror.workspace = true
+
+[dev-dependencies]
+ctor = "0.2.4"
+env_logger = "0.10"
+paste = "^1.0"
+rstest = "0.18"
