+++ crates/datafusion_ext/src/functions.rs
@@ -1,8 +1,64 @@
+use crate::errors::{ExtensionError, Result};
+
+use crate::vars::SessionVars;
+use async_trait::async_trait;
+use datafusion::arrow::datatypes::Field;
+
+use datafusion::common::OwnedTableReference;
+
+use datafusion::datasource::DefaultTableSource;
+
+use datafusion::logical_expr::{LogicalPlan, LogicalPlanBuilder};
+
+use datafusion::{arrow::datatypes::DataType, datasource::TableProvider};
+
+use std::collections::HashMap;
+use std::fmt::Write;
+
+use std::sync::Arc;
+
+use metastore_client::types::catalog::{CredentialsEntry, DatabaseEntry};
+
+#[async_trait]
+pub trait TableFunc: Sync + Send {
+    /// The name for this table function. This name will be used when looking up
+    /// function implementations.
+    fn name(&self) -> &str;
+
+    /// Return a table provider using the provided args.
+    async fn create_provider(
+        &self,
+        ctx: &dyn TableFuncContextProvider,
+        args: Vec<FuncParamValue>,
+        opts: HashMap<String, FuncParamValue>,
+    ) -> Result<Arc<dyn TableProvider>>;
+
+    /// Return a logical plan using the provided args.
+    async fn create_logical_plan(
+        &self,
+        table_ref: OwnedTableReference,
+        ctx: &dyn TableFuncContextProvider,
+        args: Vec<FuncParamValue>,
+        opts: HashMap<String, FuncParamValue>,
+    ) -> Result<LogicalPlan> {
+        let provider = self.create_provider(ctx, args, opts).await?;
+        let source = Arc::new(DefaultTableSource::new(provider));
+
+        let plan_builder = LogicalPlanBuilder::scan(table_ref, source, None)?;
+        let plan = plan_builder.build()?;
+        Ok(plan)
+    }
+}
+pub trait TableFuncContextProvider: Sync + Send {
+    fn get_database_entry(&self, name: &str) -> Option<&DatabaseEntry>;
+    fn get_credentials_entry(&self, name: &str) -> Option<&CredentialsEntry>;
+    fn get_session_vars(&self) -> &SessionVars;
+}
+
 use std::fmt;
 
 use datafusion::scalar::ScalarValue;
 
-use super::*;
 /// Value from a function parameter.
 #[derive(Debug, Clone)]
 pub enum FuncParamValue {
@@ -38,7 +94,7 @@ impl FuncParamValue {
     }
 
     /// Wrapper over `FromFuncParamValue::from_param`.
-    pub(super) fn param_into<T>(self) -> Result<T>
+    pub fn param_into<T>(self) -> Result<T>
     where
         T: FromFuncParamValue,
     {
@@ -46,7 +102,7 @@ impl FuncParamValue {
     }
 }
 
-pub(super) trait FromFuncParamValue: Sized {
+pub trait FromFuncParamValue: Sized {
     /// Get the value from parameter.
     fn from_param(value: FuncParamValue) -> Result<Self>;
 
@@ -61,7 +117,7 @@ impl FromFuncParamValue for String {
     fn from_param(value: FuncParamValue) -> Result<Self> {
         match value {
             FuncParamValue::Scalar(ScalarValue::Utf8(Some(s))) => Ok(s),
-            other => Err(BuiltinError::UnexpectedArg {
+            other => Err(ExtensionError::UnexpectedArg {
                 param: other,
                 expected: DataType::Utf8,
             }),
@@ -86,7 +142,8 @@ where
                 }
                 Ok(res)
             }
-            other => Err(BuiltinError::UnexpectedArg {
+
+            other => Err(ExtensionError::UnexpectedArg {
                 param: other,
                 expected: DataType::List(Arc::new(Field::new("list", DataType::Null, false))),
             }),
@@ -102,10 +159,10 @@ where
     }
 }
 
-pub(super) struct IdentValue(String);
+pub struct IdentValue(String);
 
 impl IdentValue {
-    pub(super) fn as_str(&self) -> &str {
+    pub fn as_str(&self) -> &str {
         self.0.as_str()
     }
 }
@@ -126,7 +183,7 @@ impl FromFuncParamValue for IdentValue {
     fn from_param(value: FuncParamValue) -> Result<Self> {
         match value {
             FuncParamValue::Ident(s) => Ok(Self(s)),
-            _other => Err(BuiltinError::Static("expected an identifier")),
+            _ => Err(ExtensionError::Static("expected an identifier")),
         }
     }
 
@@ -147,12 +204,13 @@ impl FromFuncParamValue for i64 {
                 ScalarValue::UInt16(Some(v)) => Ok(v as i64),
                 ScalarValue::UInt32(Some(v)) => Ok(v as i64),
                 ScalarValue::UInt64(Some(v)) => Ok(v as i64), // TODO: Handle overflow?
-                other => Err(BuiltinError::UnexpectedArg {
+                other => Err(ExtensionError::UnexpectedArg {
                     param: other.into(),
                     expected: DataType::Int64,
                 }),
             },
-            other => Err(BuiltinError::UnexpectedArg {
+
+            other => Err(ExtensionError::UnexpectedArg {
                 param: other,
                 expected: DataType::Int64,
             }),
@@ -188,12 +246,12 @@ impl FromFuncParamValue for f64 {
                 ScalarValue::UInt64(Some(v)) => Ok(v as f64),
                 ScalarValue::Float32(Some(v)) => Ok(v as f64),
                 ScalarValue::Float64(Some(v)) => Ok(v),
-                other => Err(BuiltinError::UnexpectedArg {
+                other => Err(ExtensionError::UnexpectedArg {
                     param: other.into(),
                     expected: DataType::Float64,
                 }),
             },
-            other => Err(BuiltinError::UnexpectedArg {
+            other => Err(ExtensionError::UnexpectedArg {
                 param: other,
                 expected: DataType::Float64,
             }),
