+++ crates/datafusion_ext/src/lib.rs
@@ -1,1 +1,6 @@
+pub mod errors;
+pub mod planner;
 pub mod vars;
+
+pub use planner::*;
+pub mod functions;
