--- crates/datafusion_planner/Cargo.toml
@@ -1,38 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-[package]
-name = "datafusion_planner"
-description = "DataFusion SQL Query Planner"
-version = { workspace = true }
-edition = { workspace = true }
-
-[features]
-default = ["unicode_expressions"]
-unicode_expressions = []
-
-[dependencies]
-datafusion = { workspace = true }
-async-trait = "0.1.72"
-async-recursion = "1.0.4"
-sqlbuiltins = {path = "../sqlbuiltins"}
-
-[dev-dependencies]
-ctor = "0.2.4"
-env_logger = "0.10"
-paste = "^1.0"
-rstest = "0.18"
