+++ crates/sqlbuiltins/src/functions.rs
@@ -1,7 +1,6 @@
 //! Builtin table returning functions.
 //! mod bigquery;
 mod bigquery;
-mod conversion;
 mod delta;
 mod generate_series;
 mod list_schemas;
@@ -24,8 +23,7 @@ use self::object_store::*;
 use self::postgres::*;
 use self::snowflake::*;
 pub(self) use self::utils::*;
-
-pub use self::conversion::*;
+pub(self) use datafusion_ext::functions::*;
 
 use crate::errors::{BuiltinError, Result};
 use async_trait::async_trait;
@@ -40,7 +38,7 @@ use datafusion::execution::TaskContext;
 use datafusion::logical_expr::{LogicalPlan, LogicalPlanBuilder};
 use datafusion::physical_plan::streaming::PartitionStream;
 use datafusion::physical_plan::{RecordBatchStream, SendableRecordBatchStream};
-use datafusion_ext::vars::SessionVars;
+use datafusion_ext::functions::TableFunc;
 
 use datafusion::{arrow::datatypes::DataType, datasource::TableProvider};
 use datasources::bigquery::{BigQueryAccessor, BigQueryTableAccess};
@@ -57,7 +55,6 @@ use datasources::object_store::{FileType, TableAccessor};
 use datasources::postgres::{PostgresAccessor, PostgresTableAccess};
 use datasources::snowflake::{SnowflakeAccessor, SnowflakeDbConnection, SnowflakeTableAccess};
 use futures::Stream;
-use metastore_client::types::catalog::{CredentialsEntry, DatabaseEntry};
 use metastore_client::types::options::{
     CredentialsOptions, DatabaseOptions, DatabaseOptionsBigQuery, DatabaseOptionsMongo,
     DatabaseOptionsMysql, DatabaseOptionsPostgres, DatabaseOptionsSnowflake,
@@ -65,7 +62,6 @@ use metastore_client::types::options::{
 use num_traits::Zero;
 use once_cell::sync::Lazy;
 use std::collections::HashMap;
-use std::fmt::Write;
 use std::ops::{Add, AddAssign};
 use std::pin::Pin;
 use std::sync::Arc;
@@ -73,37 +69,6 @@ use std::task::{Context, Poll};
 
 /// Builtin table returning functions available for all sessions.
 pub static BUILTIN_TABLE_FUNCS: Lazy<BuiltinTableFuncs> = Lazy::new(BuiltinTableFuncs::new);
-
-#[async_trait]
-pub trait TableFunc: Sync + Send {
-    /// The name for this table function. This name will be used when looking up
-    /// function implementations.
-    fn name(&self) -> &str;
-
-    /// Return a table provider using the provided args.
-    async fn create_provider(
-        &self,
-        ctx: &dyn TableFuncContextProvider,
-        args: Vec<FuncParamValue>,
-        opts: HashMap<String, FuncParamValue>,
-    ) -> Result<Arc<dyn TableProvider>>;
-
-    /// Return a logical plan using the provided args.
-    async fn create_logical_plan(
-        &self,
-        table_ref: OwnedTableReference,
-        ctx: &dyn TableFuncContextProvider,
-        args: Vec<FuncParamValue>,
-        opts: HashMap<String, FuncParamValue>,
-    ) -> Result<LogicalPlan> {
-        let provider = self.create_provider(ctx, args, opts).await?;
-        let source = Arc::new(DefaultTableSource::new(provider));
-
-        let plan_builder = LogicalPlanBuilder::scan(table_ref, source, None)?;
-        let plan = plan_builder.build()?;
-        Ok(plan)
-    }
-}
 
 /// All builtin table functions.
 pub struct BuiltinTableFuncs {
@@ -150,10 +115,4 @@ impl Default for BuiltinTableFuncs {
     fn default() -> Self {
         Self::new()
     }
-}
-
-pub trait TableFuncContextProvider: Sync + Send {
-    fn get_database_entry(&self, name: &str) -> Option<&DatabaseEntry>;
-    fn get_credentials_entry(&self, name: &str) -> Option<&CredentialsEntry>;
-    fn get_session_vars(&self) -> &SessionVars;
 }
