+++ crates/sqlbuiltins/src/functions/object_store.rs
@@ -85,7 +85,8 @@ async fn create_provider_for_filetype(
     mut args: vec::IntoIter<FuncParamValue>,
     mut opts: HashMap<String, FuncParamValue>,
 ) -> Result<Arc<dyn TableProvider>> {
-    let source_url = DatasourceUrl::new(&url_string)?;
+    let source_url =
+        DatasourceUrl::new(&url_string).map_err(|e| BuiltinError::Access(Box::new(e)))?;
 
     let provider = match args.len() {
         0 => {
@@ -129,7 +130,8 @@ async fn create_provider_for_filetype(
         }
         1 => {
             // Credentials object
-            let source_url = DatasourceUrl::new(url_string)?;
+            let source_url =
+                DatasourceUrl::new(url_string).map_err(|e| BuiltinError::Access(Box::new(e)))?;
             let creds: IdentValue = args.next().unwrap().param_into()?;
             let creds = ctx
                 .get_credentials_entry(creds.as_str())
