+++ crates/sqlexec/Cargo.toml
@@ -1,20 +1,19 @@
 [package]
 name = "sqlexec"
-version = {workspace = true}
-edition = {workspace = true}
+version = { workspace = true }
+edition = { workspace = true }
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-logutil = {path = "../logutil"}
-pgrepr = {path = "../pgrepr"}
+logutil = { path = "../logutil" }
+pgrepr = { path = "../pgrepr" }
 metastore_client = { path = "../metastore_client" }
-datafusion_planner = { path = "../datafusion_planner" }
-telemetry = {path = "../telemetry"}
-sqlbuiltins = {path = "../sqlbuiltins"}
-datasources = {path = "../datasources"}
-datafusion_ext = {path = "../datafusion_ext"}
-object_store_util = {path = "../object_store_util"}
+telemetry = { path = "../telemetry" }
+sqlbuiltins = { path = "../sqlbuiltins" }
+datasources = { path = "../datasources" }
+datafusion_ext = { path = "../datafusion_ext" }
+object_store_util = { path = "../object_store_util" }
 thiserror.workspace = true
 tokio = { version = "1", features = ["full"] }
 async-trait = "0.1.72"
