+++ crates/sqlexec/src/planner/context_builder.rs
@@ -15,14 +15,14 @@ use datafusion::logical_expr::AggregateUDF;
 use datafusion::logical_expr::ScalarUDF;
 use datafusion::logical_expr::TableSource;
 use datafusion::sql::TableReference;
+use datafusion_ext::functions::TableFunc;
+use datafusion_ext::functions::TableFuncContextProvider;
+use datafusion_ext::planner::AsyncContextProvider;
 use datafusion_ext::vars::SessionVars;
-use datafusion_planner::planner::AsyncContextProvider;
 use metastore_client::types::catalog::CatalogEntry;
 use metastore_client::types::catalog::CredentialsEntry;
 use metastore_client::types::catalog::DatabaseEntry;
 use sqlbuiltins::builtins::DEFAULT_CATALOG;
-use sqlbuiltins::functions::TableFunc;
-use sqlbuiltins::functions::TableFuncContextProvider;
 use sqlbuiltins::functions::BUILTIN_TABLE_FUNCS;
 use std::collections::HashMap;
 use std::sync::Arc;
